% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.files.R
\name{commit.files}
\alias{commit.files}
\title{Commit files to the "incoming" project subfolder on the USFWS Alaska Regional Data Repository (RDR)}
\usage{
commit.files(project, local.folder, recursive)
}
\arguments{
\item{project}{Character string. Name of the project folder.}

\item{local.folder}{Character string. Directory name or path where the uncommitted files are located. Default is the working directory, getwd().}

\item{recursive}{Logical. Whether to search for and commit files in subdirectories. Default is TRUE.}
}
\value{
Returns a vector of the committed files.
}
\description{
Copies local file(s) to the "incoming" subfolder of a specified RDR project and updates the changelog to document changes. Remote users must be connected the Service’s approved remote connection technologies, such as a Virtual Private Network (VPN) AND be granted write permission to the project’s “incoming" subfolder (i.e., be an authorized "data steward").
}
\examples{
\dontrun{
e.g.commit<- commit.files(project = "mbmlb_010_Grey_headed_chickadee_hybridization", local.folder = getwd(), recursive = TRUE)
}
}
\seealso{
\code{commit.files()}
}
\keyword{USFWS,}
\keyword{repository}
