% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.files.R
\name{download.files}
\alias{download.files}
\title{Download project file(s) from the USFWS Alaska Regional Data Repository (RDR)}
\usage{
download.files(
  pattern,
  project,
  subfolder.path,
  local.path,
  main,
  incoming,
  recursive,
  download.file.method
)
}
\arguments{
\item{pattern}{Character vector. File name pattern(s). Must be a regular expression; print ?base::regex for help. Default is NULL, which allows a selection from all files.}

\item{project}{Character string. Name of the project folder.}

\item{subfolder.path}{Character string. Project subfolder path.}

\item{local.path}{Character string. Directory path where the downloaded files will be saved. Default is the working directory, getwd().}

\item{main}{Logical. Whether to return results from the main project subfolders (all subfolders except "incoming"). Default is TRUE.}

\item{incoming}{Logical. Whether to return results from the "incoming" project subfolder. Default is TRUE.}

\item{recursive}{Logical. Whether to search for and download files in subdirectories. Default is TRUE.}

\item{download.file.method}{Character string. Method to use for downloading files. Print ?download.file for available methods. Default is "auto".}
}
\value{
Returns file download(s) which match the search criteria.
}
\description{
Searches for file name patterns in a specified RDR project folder and downloads all matching files. Remote users must be connected to one of the Service’s approved remote connection technologies, such as a Virtual Private Network (VPN).
}
\examples{
# download.files(pattern = c("template","dictionary","\\.csv","hello"), project = "mbmlb_007_NWR_Alaska_Landbird_Monitoring_Survey", subfolder.path = "metadata, local.path = getwd(), main = TRUE, incoming = TRUE, recursive = TRUE, download.file.method = "curl")
}
\seealso{
\code{read.tables()}
}
\keyword{USFWS,}
\keyword{download,}
\keyword{files}
\keyword{repository,}
