% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.files.R
\name{find.files}
\alias{find.files}
\title{Find the location(s) of project file(s) on the USFWS Alaska Regional Data Repository (RDR)}
\usage{
find.files(
  pattern,
  project,
  subfolder.path,
  main,
  incoming,
  recursive,
  full.path
)
}
\arguments{
\item{pattern}{Character vector. File name pattern(s). Must be a regular expression; print ?base::regex for help. Not case-sensitive. Default is NULL, which returns results for all files.}

\item{project}{Character string. Project folder name. Can be a partial name formatted as a regular expression. Not case-sensitive.}

\item{subfolder.path}{Character string. Project subfolder path.}

\item{main}{Logical. Whether to return results from the main project subfolders (all subfolders except "incoming"). Default is TRUE.}

\item{incoming}{Logical. whether to return results from the "incoming" project subfolder. Default is TRUE.}

\item{recursive}{Logical. Whether to search for files in subdirectories. Default is TRUE.}

\item{full.path}{Logical. Whether to return full file path. Default is FALSE}
}
\value{
Returns a vector of paths to files which match the search criteria.
}
\description{
Finds data file(s) from a specified RDR project folder. Remote users must be connected to one of the Service’s approved remote connection technologies, such as a Virtual Private Network (VPN).
}
\examples{
# e.g.file.locs<- find.files(pattern = c("template","dictionary","\\.csv","hello"), project = "mbmlb_007_NWR_Alaska_Landbird_Monitoring_Survey", subfolder.path = "", main = FALSE, incoming = TRUE, recursive = TRUE, full.path = FALSE)
}
\seealso{
\code{download.files()}
}
\keyword{USFWS,}
\keyword{repository}
