% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.contacts.R
\name{get.contacts}
\alias{get.contacts}
\title{Get contacts from an mdEditor metadata file}
\usage{
get.contacts(x)
}
\arguments{
\item{x}{An mdEditor metadata file containing one or more contact records}
}
\value{
Returns a data frame summarizing all contacts found within the metadata file, including the contactId, name, and email address.
}
\description{
Generates a table of contacts for individuals and organizations associated with a set of metadata records.
}
\examples{
# get.contacts(file = "AK-contacts-mdeditor-20241210-111268.json")
}
