% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contacts.R
\name{get_contacts}
\alias{get_contacts}
\title{Get Contacts}
\usage{
get_contacts(file)
}
\arguments{
\item{file}{A mdJSON metadata file containing a list of contacts}
}
\value{
A data.frame containing all contacts found within metadata file,
including their contactIds, names, and email addresses.
}
\description{
Generates a table of contacts for individuals and organizations associated
with a set of metadata records. Primarily used within other functions.
}
