% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.mdEditor.R
\name{scrape.mdEditor}
\alias{scrape.mdEditor}
\title{Scrape mdEditor metadata}
\usage{
scrape.mdEditor(x)
}
\arguments{
\item{x}{An mdEditor metadata file containing one or more record. mdJSON files are not compatible.}
}
\value{
Returns a data frame summarizing the file name and select metadata record parameters. Each row is a unique record.
}
\description{
Extracts select metadata parameters from an mdEditor file into a dataframe.
}
\examples{
# scrape.mdEditor(x = "fescgl_001_MixedStockAnalysis-mdeditor-20240805-110842.json")
}
