% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_mdEditor.R
\name{scrape_mdEditor}
\alias{scrape_mdEditor}
\title{Scrape mdEditor metadata}
\usage{
scrape_mdEditor(x)
}
\arguments{
\item{x}{mdEditor produced JSON file}
}
\value{
A dataframe containing the file name and select metadata record
parameters. Each row is a unique record.
}
\description{
Extracts select metadata parameters from mdEditor file into a dataframe.
Note that only mdeditor files, not mdjson files, are not compatible with this
function.
}
\examples{
# scrape_mdEditor(x = "fescgl_001_MixedStockAnalysis-mdeditor-20240805-110842.json")
}
